Partially revert "Merge pull request #134 from smartsheet/issue-133-decimal-parsing-bug"
This reverts commit a3f392c13d569c48af5fe9531194bdc79f777342, reversing
changes made to 4a833d060a23409fcbda74053fbc891c68d009ba.

+++ smartsheet-csharp-sdk/main/Smartsheet/Api/Internal/Json/ObjectValueTypeConverter.cs
@@ -6,9 +6,9 @@
 //    Licensed under the Apache License, Version 2.0 (the "License");
 //    you may not use this file except in compliance with the License.
 //    You may obtain a copy of the License at
-//        
+//
 //            http://www.apache.org/licenses/LICENSE-2.0
-//        
+//
 //    Unless required by applicable law or agreed to in writing, software
 //    distributed under the License is distributed on an "AS IS" BASIS,
 //    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -56,9 +56,9 @@ namespace Smartsheet.Api.Internal.Json
                 ObjectValueAttributeSuperset superset = new ObjectValueAttributeSuperset();
                 serializer.Populate(reader, superset);
                 ObjectValueType parsedObjectType;
-                if(!Enum.TryParse(superset.objectType, true, out parsedObjectType)) 
+                if(!Enum.TryParse(superset.objectType, true, out parsedObjectType))
                 {
-                    // If a new object type is introduced to the Smartsheet API that this version of the SDK doesn't support, 
+                    // If a new object type is introduced to the Smartsheet API that this version of the SDK doesn't support,
                     // return null instead of throwing an exception.
                     return null;
                 }
@@ -117,11 +117,11 @@ namespace Smartsheet.Api.Internal.Json
                 }
                 else if (reader.TokenType == JsonToken.Integer)
                 {
-                    objectValue = new NumberObjectValue(Convert.ToDecimal(reader.Value));
+                    objectValue = new NumberObjectValue(Convert.ToDouble(reader.Value));
                 }
                 else if (reader.TokenType == JsonToken.Float)
                 {
-                    objectValue = new NumberObjectValue((decimal)reader.Value);
+                    objectValue = new NumberObjectValue((double)reader.Value);
                 }
                 else if (reader.TokenType == JsonToken.Date)
                 {
